+++ research/lance-autoresearch/crates/pq-l2/src/kernels.rs
